HHH-12052 - Move PooledConnections inside DriverManagerConnectionProviderImpl
